backend-server/vision-edit_fastapi/requirements.txt
fastapi==0.115.12
uvicorn==0.33.0
sqlalchemy==1.4.15
pymysql==1.1.1
sentence-transformers==3.0.1
faiss-cpu>=1.9.0
numpy>=1.24.0
pandas>=1.3.4
python-dotenv>=1.0.0
# M1 Mac (Apple Silicon) 호환 PyTorch - CPU/MPS 지원
torch>=2.0.0
httpx>=0.24.0
langchain==0.3.27
langchain-community==0.3.16
transformers>=4.35.0
accelerate>=0.25.0
ultralytics>=8.2.0
python-multipart>=0.0.9
